mcp-poc/services/context-service/requirements.txt
# Core Python packages (pinned for stability)
pip==23.3.1
setuptools==69.0.3
wheel==0.42.0

# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
asyncpg==0.29.0
sqlalchemy==2.0.23

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.8
httpx==0.25.2

# Configuration and validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Logging and observability
structlog==23.2.0
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-httpx==0.42b0
opentelemetry-instrumentation-asyncpg==0.42b0
opentelemetry-exporter-otlp-proto-grpc==1.21.0

# CORS and middleware
python-cors==1.7.0

# Database migrations
alembic==1.13.1

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
